@@ -10,7 +10,7 @@ import { HeaderComponent } from '../components/header/header.component'
 import { PriestsComponent } from '../components/priests/priests.component'
 import { ContactComponent } from '../components/contact/contact.component'
 import { FooterComponent } from '../components/footer/footer.component'
-import { HttpClient, HttpClientModule } from '@angular/common/http'
+import { HttpClient } from '@angular/common/http'
 
 @Component({
   selector: 'app-hawick-home',
@@ -25,7 +25,11 @@ import { HttpClient, HttpClientModule } from '@angular/common/http'
     PriestsComponent,
     ContactComponent,
     FooterComponent,
-    HttpClientModule
+    
+// TODO: `HttpClientModule` should not be imported into a component directly.
+// Please refactor the code to add `provideHttpClient()` call to the provider list in the
+// application bootstrap logic and remove the `HttpClientModule` import from this component.
+HttpClientModule
   ],
   templateUrl: './home.component.html',
   styleUrl: './home.component.scss'
